refactor: rename package to handler follow go convention
