+++ internal/handler/leasing_history_handler_create.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"github.com/PitiNarak/condormhub-backend/internal/dto"
