+++ internal/handler/user_handler_delete_account.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"github.com/PitiNarak/condormhub-backend/pkg/apperror"
