+++ internal/handler/user_handler_get_user_info.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
