+++ internal/handler/user_handler_reset_password.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"errors"
