+++ internal/handler/user_handler_update_user_information.go
@@ -1,4 +1,4 @@
-package handlers
+package handler
 
 import (
 	"errors"
