+++ internal/server/handler.go
@@ -2,13 +2,13 @@ package server
 
 import (
 	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
-	"github.com/PitiNarak/condormhub-backend/internal/handlers"
+	handler1 "github.com/PitiNarak/condormhub-backend/internal/handler"
 )
 
 type handler struct {
-	greeting       *handlers.GreetingHandler
+	greeting       *handler1.GreetingHandler
 	user           ports.UserHandler
-	exampleUpload  *handlers.TestUploadHandler
+	exampleUpload  *handler1.TestUploadHandler
 	dorm           ports.DormHandler
 	leasingHistory ports.LeasingHistoryHandler
 	order          ports.OrderHandler
@@ -16,13 +16,13 @@ type handler struct {
 }
 
 func (s *Server) initHandler() {
-	greeting := handlers.NewGreetingHandler()
-	user := handlers.NewUserHandler(s.service.user)
-	exampleUpload := handlers.NewTestUploadHandler(s.storage)
-	dorm := handlers.NewDormHandler(s.service.dorm)
-	leasingHistory := handlers.NewLeasingHistoryHandler(s.service.leasingHistory)
-	order := handlers.NewOrderHandler(s.service.order)
-	tsx := handlers.NewTransactionHandler(s.service.tsx, s.stripeConfig)
+	greeting := handler1.NewGreetingHandler()
+	user := handler1.NewUserHandler(s.service.user)
+	exampleUpload := handler1.NewTestUploadHandler(s.storage)
+	dorm := handler1.NewDormHandler(s.service.dorm)
+	leasingHistory := handler1.NewLeasingHistoryHandler(s.service.leasingHistory)
+	order := handler1.NewOrderHandler(s.service.order)
+	tsx := handler1.NewTransactionHandler(s.service.tsx, s.stripeConfig)
 
 	s.handler = &handler{
 		greeting:       greeting,
